Begin to work on reworking theme.slice functions [ci skip]
