+++ src/data/constants.ts
@@ -46,7 +46,7 @@ export const CORNER_INDEX_TO_POSITIONING = [
 ];
 
 // Local storage keys
-export const STORED_PREFERENCE_KEY = "stored_color_preference_v2";
+export const STORED_IGNORE_COMPUTER_PREFERENCE = "stored_color_preference_v2";
 export const STORED_THEMES = "stored_themes_v2";
-export const STORED_PREFERENCES = "stored_preferences_v2";
-export const LAST_USED_THEME = "last_used_theme_v2";
+export const STORED_ACTIVE_THEME = "stored_active_theme_v2";
+export const STORED_LAST_USED_THEME_ID = "last_used_theme_v2";
