+++ src/store/theme/theme.slice.ts
@@ -1,57 +1,40 @@
 import { type PayloadAction, createSlice } from "@reduxjs/toolkit";
 
-import { defaultDayTheme, initialState } from "./theme.state";
-import { flattenTheme } from "@/utils/other";
+import { initialState } from "./theme.state";
 import { StringLookup } from "@/types/general";
 import {
-	STORED_PREFERENCE_KEY,
-	STORED_PREFERENCES,
+	STORED_IGNORE_COMPUTER_PREFERENCE,
+	STORED_ACTIVE_THEME,
 	STORED_THEMES,
+	STORED_LAST_USED_THEME_ID,
 } from "@/data/constants";
-
-function determineComputerPreferredTheme(state: ThemeState) {
-	const darkThemeTime = window.matchMedia(
-		"(prefers-color-scheme: dark)",
-	).matches;
-	const themePreference = darkThemeTime ? "1" : "0";
-	if (state.active.id !== themePreference) {
-		return (
-			state.themes.find((theme) => theme.id === themePreference) ?? state.active
-		);
-	}
-	return state.active;
-}
+import {
+	determineComputerPreferredTheme,
+	defaultDayTheme,
+	getDefaultThemeState,
+} from "./theme.utils";
 
 function copyTheme(
 	copiedTheme: BaseTheme,
 	state: ThemeState,
 ): { name: string; id: number } {
 	let { name } = copiedTheme;
-	let nameIndex = 1;
 	const match = name.match(/\d+$/);
 	if (match) {
 		const digits = match[0];
 		name = name.slice(0, -digits.length);
-		nameIndex = +digits + 1;
 	}
 
-	let finalName = `${name}${nameIndex}`;
-
-	while (state.themes.find((theme) => theme.name === finalName)) {
-		if (nameIndex === Number.MAX_SAFE_INTEGER) {
-			nameIndex = 0;
-		}
-
-		if (nameIndex > 0) {
-			nameIndex++;
-		} else {
-			nameIndex--;
-		}
-
-		finalName = `${name}${nameIndex}`;
+	let themeId = state.latestId;
+	if (themeId >= Number.MAX_SAFE_INTEGER) {
+		themeId = -1;
+	} else if (themeId < 0) {
+		themeId--;
+	} else {
+		themeId++;
 	}
 
-	return { name: finalName, id: state.latestId + 1 };
+	return { name: `${name}${themeId}`, id: themeId };
 }
 
 const themeSlice = createSlice({
@@ -65,6 +48,7 @@ const themeSlice = createSlice({
 					? state.themes.find((theme) => theme.id === "1")
 					: state.themes.find((theme) => theme.id === "0");
 			state.active = newTheme ? newTheme : state.active;
+			localStorage.setItem(STORED_ACTIVE_THEME, state.active.id);
 		},
 
 		setActiveThemeByID: (state, action: PayloadAction<string>) => {
@@ -72,17 +56,19 @@ const themeSlice = createSlice({
 				(theme) => theme.id === action.payload,
 			);
 			state.active = newTheme ? newTheme : state.active;
+			localStorage.setItem(STORED_ACTIVE_THEME, state.active.id);
 		},
 
 		toggleUseComputerPreferences: (state) => {
 			state.ignoreComputerPreferences = !state.ignoreComputerPreferences;
-			localStorage.setItem(
-				STORED_PREFERENCE_KEY,
-				state.ignoreComputerPreferences.toString(),
-			);
 			if (!state.ignoreComputerPreferences) {
 				state.active = determineComputerPreferredTheme(state);
 			}
+
+			localStorage.setItem(
+				STORED_IGNORE_COMPUTER_PREFERENCE,
+				state.ignoreComputerPreferences.toString(),
+			);
 		},
 
 		copyThemeByID: (state, action: PayloadAction<string>) => {
@@ -95,16 +81,21 @@ const themeSlice = createSlice({
 			}
 
 			const { name, id } = copyTheme(copiedTheme, state);
-			// @ts-ignore
-			state.themes.push({ ...copiedTheme, name, id });
-			state.latestId = +id;
+			state.themes.push({ ...copiedTheme, name, id: id.toString() });
+			state.latestId = id;
 			localStorage.setItem(STORED_THEMES, JSON.stringify(state.themes));
+			localStorage.setItem(STORED_LAST_USED_THEME_ID, id.toString());
 		},
 
 		createTheme: (state) => {
 			const { name, id } = copyTheme(defaultDayTheme, state);
 			state.latestId = id;
 			state.themes.push({ ...defaultDayTheme, name, id: id.toString() });
+
+			localStorage.setItem(
+				STORED_LAST_USED_THEME_ID,
+				state.latestId.toString(),
+			);
 			localStorage.setItem(STORED_THEMES, JSON.stringify(state.themes));
 		},
 
@@ -121,6 +112,7 @@ const themeSlice = createSlice({
 			}
 
 			state.themes[themeToUpdateIndex] = theme;
+
 			localStorage.setItem(STORED_THEMES, JSON.stringify(state.themes));
 		},
 
@@ -135,11 +127,13 @@ const themeSlice = createSlice({
 
 			if (state.active.id === action.payload) {
 				state.active = state.themes[themeIndexToDelete === 0 ? 1 : 0];
+				localStorage.setItem(STORED_ACTIVE_THEME, state.active.id);
 			}
 
 			state.themes = state.themes.filter(
 				(theme) => theme.id !== action.payload,
 			);
+
 			localStorage.setItem(STORED_THEMES, JSON.stringify(state.themes));
 		},
 
@@ -163,6 +157,7 @@ const themeSlice = createSlice({
 				state.active =
 					state.themes.find((theme) => theme.id === id) ?? state.active;
 			}
+
 			localStorage.setItem(STORED_THEMES, JSON.stringify(state.themes));
 		},
 
@@ -173,7 +168,7 @@ const themeSlice = createSlice({
 			const { id, props, newVal } = action.payload;
 			const accessor = state.themes.find((theme) => theme.id === id);
 
-			if (!accessor || accessor.id === "0" || accessor.id === "1") {
+			if (!accessor) {
 				return;
 			}
 
@@ -184,6 +179,7 @@ const themeSlice = createSlice({
 				if (accessor.length === 1) {
 					return obj[accessor[0]] as StringLookup;
 				}
+
 				if (obj[accessor[0]]) {
 					return recursiveAccess(
 						obj[accessor[0]] as RecursiveControlGroup,
@@ -192,6 +188,7 @@ const themeSlice = createSlice({
 				}
 				return;
 			}
+
 			const finalAccessor = recursiveAccess(accessor, props.slice(0, -1));
 			if (!finalAccessor) {
 				return;
@@ -203,34 +200,17 @@ const themeSlice = createSlice({
 				state.active =
 					state.themes.find((theme) => theme.id === id) ?? state.active;
 			}
+
 			localStorage.setItem(STORED_THEMES, JSON.stringify(state.themes));
 		},
-		intializeThemeStore: (
-			state,
-			action: PayloadAction<{
-				computerPreferences: boolean;
-				themes: string | null;
-				preference: string | null | undefined;
-			}>,
-		) => {
-			state.ignoreComputerPreferences = action.payload.computerPreferences;
-			state.themes = action.payload.themes
-				? (JSON.parse(action.payload.themes) as BaseTheme[])
-				: state.themes;
-			// Find the last active theme from local storage
-			const activeTheme = state.themes.find(
-				// @ts-ignore
-				(theme) => theme.id === state.prefers,
-			);
-			state.active = activeTheme ?? state.active;
-		},
+
 		resetThemeOptions: () => {
 			localStorage.removeItem(STORED_THEMES);
-			localStorage.removeItem(STORED_PREFERENCES);
-			localStorage.removeItem(STORED_PREFERENCE_KEY);
-			const defaultState = { ...initialState };
-			defaultState.active = determineComputerPreferredTheme(defaultState);
-			return defaultState;
+			localStorage.removeItem(STORED_ACTIVE_THEME);
+			localStorage.removeItem(STORED_IGNORE_COMPUTER_PREFERENCE);
+			localStorage.removeItem(STORED_LAST_USED_THEME_ID);
+
+			return getDefaultThemeState();
 		},
 	},
 });
@@ -245,10 +225,7 @@ export const {
 	deleteThemeByID,
 	changeThemeName,
 	changePropOnTheme,
-	intializeThemeStore,
 	resetThemeOptions,
 } = themeSlice.actions;
 
 export default themeSlice.reducer;
-
-export const flattenedThemeShape = flattenTheme(initialState.active);
